refactor: update tsx ls-lint
