+++ packages/cdk/virtual-list/src/ScrollBar.tsx
