+++ packages/components/pagination/src/Total.tsx
