+++ packages/components/space/index.ts
@@ -1,6 +1,6 @@
 import type { App } from 'vue'
 
-import IxSpace from './src/space'
+import IxSpace from './src/Space'
 
 IxSpace.install = (app: App): void => {
   app.component(IxSpace.name, IxSpace)
